+++ AeViz/AeVizMethods/__init__.py
@@ -1,3 +1,3 @@
 from typing import Literal
-from AeViz.utils.decorators import fig_window_open
+from AeViz.utils.decorators.decorators import fig_window_open
 from AeViz.utils.aeviz_utils import plot_qt
