+++ AeViz/simulation/methods/__init__.py
@@ -1,5 +1,5 @@
 from AeViz.units import u
 from AeViz.units.aeseries import aerray, aeseries
-from AeViz.utils.decorators import (smooth, derive, hdf_isopen,
+from AeViz.utils.decorators.decorators import (smooth, derive, hdf_isopen,
                                     subtract_tob, sum_tob)
 import numpy as np
